Add Apache 2.0 License
- Include LICENSE.txt file with Apache 2.0 license text

This commit adds the Apache 2.0 license to the project, clarifying the terms
under which the project can be used, modified, and distributed. The Apache 2.0
license is a permissive free software license that allows users to use the
software for any purpose, to distribute it, to modify it, and to distribute
modified versions of the software.

Key points:
- Allows commercial use
- Allows modification and distribution
- Requires preservation of copyright and license notices
- Provides an express grant of patent rights from contributors to users
- Requires attribution for any changes made to the software